fix(request): исправить валидацию при создании

+++ ewm-main-service/src/main/resources/schema.sql
@@ -41,7 +41,7 @@ CREATE TABLE IF NOT EXISTS events (
     request_moderation  BOOLEAN                                 NOT NULL,
     state               VARCHAR(20)                             NOT NULL,
     title               VARCHAR(120)                            NOT NULL,
-    views               INTEGER                                 NOT NULL,
+    views               BIGINT                                  NOT NULL,
 
     CONSTRAINT events_p_key        PRIMARY KEY (id),
     CONSTRAINT events_category_fk  FOREIGN KEY (category)   REFERENCES categories (id) ON DELETE CASCADE,
